@@ -1,23 +1,17 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
-
-	<PropertyGroup>
-		<TargetFramework>net6.0</TargetFramework>
-		<ImplicitUsings>enable</ImplicitUsings>
-		<OutputType>Library</OutputType>
-	</PropertyGroup>
-
-	<ItemGroup>
-		<ProjectReference Include="..\Pixel.Identity.Shared\Pixel.Identity.Shared.csproj" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<PackageReference Include="AutoMapper" Version="12.0.1" />		
-		<PackageReference Include="OpenIddict.AspNetCore" Version="4.7.0" />
-	</ItemGroup>
-
-
-	<ItemGroup>
-		<Using Include="Pixel.Identity.Core.Extensions" />	
-	</ItemGroup>
-
+  <PropertyGroup>
+    <TargetFramework>net8.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <OutputType>Library</OutputType>
+  </PropertyGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\Pixel.Identity.Shared\Pixel.Identity.Shared.csproj" />
+  </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="AutoMapper" Version="12.0.1" />
+    <PackageReference Include="OpenIddict.AspNetCore" Version="5.0.1" />
+  </ItemGroup>
+  <ItemGroup>
+    <Using Include="Pixel.Identity.Core.Extensions" />
+  </ItemGroup>
 </Project>
