@@ -1,33 +1,29 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
-
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>   
-    <ImplicitUsings>enable</ImplicitUsings>   
+    <TargetFramework>net8.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
     <OutputType>Library</OutputType>
-	<EnableDynamicLoading>true</EnableDynamicLoading>
+    <EnableDynamicLoading>true</EnableDynamicLoading>
   </PropertyGroup>
-
-	<ItemGroup>		
-		<PackageReference Include="AutoMapper" Version="12.0.1">
-			<ExcludeAssets>runtime</ExcludeAssets>
-		</PackageReference>		
-		<PackageReference Include="Dawn.Guard" Version="1.12.0" />	
-		<PackageReference Include="MongoDB.Driver" Version="2.21.0" />
-		<PackageReference Include="OpenIddict.Abstractions" Version="4.7.0">
-			<ExcludeAssets>runtime</ExcludeAssets>
-		</PackageReference>
-		<PackageReference Include="OpenIddict.Quartz" Version="4.7.0">
-			<ExcludeAssets>runtime</ExcludeAssets>
-		</PackageReference>
-		<PackageReference Include="OpenIddict.MongoDb" Version="4.7.0" />
-		<PackageReference Include="OpenIddict.MongoDb.Models" Version="4.7.0" />		
-	</ItemGroup>
-
-	<ItemGroup>
-		<ProjectReference Include="..\Pixel.Identity.Core\Pixel.Identity.Core.csproj">
-			<Private>false</Private>
-			<ExcludeAssets>runtime</ExcludeAssets>
-		</ProjectReference>	
-	</ItemGroup>
-	
+  <ItemGroup>
+    <PackageReference Include="AutoMapper" Version="12.0.1">
+      <ExcludeAssets>runtime</ExcludeAssets>
+    </PackageReference>
+    <PackageReference Include="Dawn.Guard" Version="1.12.0" />
+    <PackageReference Include="MongoDB.Driver" Version="2.23.1" />
+    <PackageReference Include="OpenIddict.Abstractions" Version="5.0.1">
+      <ExcludeAssets>runtime</ExcludeAssets>
+    </PackageReference>
+    <PackageReference Include="OpenIddict.Quartz" Version="5.0.1">
+      <ExcludeAssets>runtime</ExcludeAssets>
+    </PackageReference>
+    <PackageReference Include="OpenIddict.MongoDb" Version="5.0.1" />
+    <PackageReference Include="OpenIddict.MongoDb.Models" Version="5.0.1" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\Pixel.Identity.Core\Pixel.Identity.Core.csproj">
+      <Private>false</Private>
+      <ExcludeAssets>runtime</ExcludeAssets>
+    </ProjectReference>
+  </ItemGroup>
 </Project>
