@@ -5,9 +5,9 @@ namespace Pixel.Identity.Shared.Request
 {
     public class RemoveUserRolesRequest
     {
-        public string UserName { get; set; }
+        public string UserName { get; set; } = string.Empty;
 
-        public List<UserRoleViewModel> RolesToRemove { get; set; } = new List<UserRoleViewModel>();
+        public List<UserRoleViewModel> RolesToRemove { get; set; } = [];
 
         public RemoveUserRolesRequest()
         {
