@@ -6,6 +6,6 @@ namespace Pixel.Identity.Shared.Models
     {
         [Required]
         [EmailAddress]
-        public string NewEmail { get; set; }
+        public string NewEmail { get; set; } = string.Empty;
     }
 }
