@@ -6,6 +6,6 @@ namespace Pixel.Identity.Shared.Models
     {
         [Required]
         [DataType(DataType.Password)]
-        public string Password { get; set; }
+        public string Password { get; set; } = string.Empty;
     }
 }
