@@ -12,14 +12,14 @@ namespace Pixel.Identity.Shared.Request
         /// </summary>
         [Required]
         [DataMember(IsRequired = true)]      
-        public string Owner { get; set; }
+        public string Owner { get; set; } = string.Empty;
 
         /// <summary>
         /// Claim to add to the role
         /// </summary>
         [Required]
         [DataMember(IsRequired = true)]
-        public ClaimViewModel ClaimToAdd { get; set; }
+        public ClaimViewModel ClaimToAdd { get; set; } = new ClaimViewModel();
 
         /// <summary>
         /// constructor
