@@ -5,6 +5,6 @@ namespace Pixel.Identity.Shared.Request
     [DataContract]
     public class GetScopesRequest : PagedDataRequest
     {
-       public string ScopesFilter { get; set; }
+       public string ScopesFilter { get; set; } = string.Empty;
     }
 }
