@@ -27,10 +27,10 @@ internal class PageSesionTest
 
     [OneTimeSetUp]
     public async Task PlaywrightSetup()
-    {
+    {       
         Playwright = await playwrightTask.ConfigureAwait(false);
         BrowserType = Playwright[BrowserName];    
-        Assert.IsNotNull(BrowserType, $"The requested browser ({BrowserName}) could not be found - make sure your BROWSER env variable is set correctly.");
+        Assert.That(BrowserType, Is.Not.Null, $"The requested browser ({BrowserName}) could not be found - make sure your BROWSER env variable is set correctly.");
         Browser = await BrowserType.LaunchAsync(new()
         {
             Headless = Environment.GetEnvironmentVariable("HEADED") != "1"
